refactor: restructure component folder structure to prevent project sync starter conflict
