+++ package.json
@@ -56,7 +56,7 @@
     "react-responsive-carousel": "^3.2.21",
     "react-slick": "^0.28.1",
     "reconnect.js": "^1.1.1",
-    "rev.sdk.js": "^0.1.52",
+    "rev.sdk.js": "^0.1.54",
     "stream-browserify": "^3.0.0",
     "stream-http": "^3.2.0",
     "styled-components": "^5.3.0",
