--- src/Components/Layout.css
@@ -1,19 +1,0 @@
-/* add custom css here */
-p,
-h1,
-h2,
-h3,
-h4,
-h5,
-h6 {
-    margin: 0;
-}
-
-@keyframes navMove {
-    from {
-        background-position: left bottom;
-    }
-    to {
-        background-position: -200px bottom;
-    }
-}
