+++ src/Components/Layout/Layout.css
@@ -1,0 +1,10 @@
+/* add custom css here */
+p,
+h1,
+h2,
+h3,
+h4,
+h5,
+h6 {
+    margin: 0;
+}
